Enh: ability to delete a single downtime
This patch bring the ability to delete a single downtime as opposed to
all the downtimes of an host or service in the current implementation,
and remains the default behavior.

This is done by simply providing the parameter `downtime_id` to the
`/downtime` endpoint when `action=delete`.

Also re-indented some code for better readability and removed Python 2.6
test from Travis configuration as this is no more available.